requirements.txt
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.9
httpx>=0.27
pydantic>=2.7
itsdangerous>=2.2
werkzeug>=3.0
jinja2>=3.1
reportlab>=4.0
python-dotenv>=1.0
openpyxl>=3.1
pywebview>=5.0
pg8000>=1.31
qrcode>=8.0
redis>=5.2
coverage>=7.6
mypy>=1.11
black>=24.8
ruff>=0.6
